+++ fuel-merkle/src/sparse/merkle_tree/node.rs
@@ -104,7 +104,8 @@ impl Node {
             // of the leaves or an ancestor multiple generations above the
             // leaves.
             // N.B.: A leaf can be a placeholder.
-            let parent_depth = path_node.common_path_length(side_node);
+            #[allow(clippy::cast_possible_truncation)] // Key is 32 bytes
+            let parent_depth = path_node.common_path_length(side_node) as u32;
             #[allow(clippy::arithmetic_side_effects)] // parent_depth <= max_height
             let parent_height = Node::max_height() - parent_depth;
             match path.get_instruction(parent_depth).unwrap() {
@@ -131,7 +132,7 @@ impl Node {
         Self::Placeholder
     }
 
-    pub fn common_path_length(&self, other: &Node) -> u32 {
+    pub fn common_path_length(&self, other: &Node) -> u64 {
         debug_assert!(self.is_leaf());
         debug_assert!(other.is_leaf());
